factory-pattern/requirements.txt
pymongo==3.7.2
mysql-connector==2.1.6